requirements.txt
# Core
Django==5.2.11
django-allauth==65.14.3
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1

# Database
PyMySQL==1.1.2

# Auth & crypto
PyJWT==2.11.0
cryptography==46.0.5

# Server & static
gunicorn==25.1.0
whitenoise==6.12.0

# Media
pillow==12.1.1

# HTTP & requests
requests==2.32.5
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
urllib3==2.6.3

# Testing
pytest==9.0.2
pytest-django==4.11.1
iniconfig==2.3.0
pluggy==1.6.0

# Config (optional - project uses os.environ; keep for .env support if needed)
python-dotenv==1.2.1

# YAML (Django serializers, DRF, allauth)
PyYAML==6.0.3

# Transitive / low-level
asgiref==3.11.1
attrs==25.4.0
cffi==2.0.0
pycparser==3.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
